Raise max TCP segment queue length
Out-of-order TCP segments are queued in the reassembly queue until the missing in-sequence segments arrive.  As a DoS mitigation, the length of this queue is limited, with a default of 100.  Since TCP allows a whole window to be in-flight at once, there can be up to 1448 to 11586 segments arriving before the sender expects an ACK, depending on the configured max receive buffer size and negotiated window size.  The limit of 100 is not usually an issue in wired networks, because segments seldom arrive out of order.  However, reordering and/or losses are more frequent with wireless networks and more complex networks, so the low default limit can cause a lot of segments to be discarded.

Raise the limit to 1448, which allows a full window to be queued in the default configuration.

See following ticket: https://redmine.ixsystems.com/issues/43558